requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
playwright>=1.40.0
pandas>=2.2.0
openpyxl>=3.1.2
Pillow>=10.0.0
pytesseract>=0.3.10
